requirements.txt
Django>=4.2
djangorestframework>=3.15
numpy>=1.26
scikit-learn>=1.4
faker>=18.3.1
